rename package pkg/audit to satellite/audit
