+++ satellite/audit/containment.go
