+++ satellite/audit/reporter.go
