+++ satellite/audit/testplanet_helpers_test.go
