+++ satellite/audit/verifier.go
