+++ satellite/audit/verifier_unit_test.go
