+++ satellite/satellitedb/containment.go
@@ -10,9 +10,9 @@ import (
 
 	"github.com/zeebo/errs"
 
-	"storj.io/storj/pkg/audit"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/storj"
+	"storj.io/storj/satellite/audit"
 	dbx "storj.io/storj/satellite/satellitedb/dbx"
 )
